--- CarRentalCompany.cpp
@@ -1,8 +1,0 @@
-#include <iostream>
-
-using namespace std;
-
-int main()
-{
-
-}
