+++ Client.h
@@ -1,7 +1,4 @@
 #include <iostream>
-
-#include <RentalData.h>
-
 using namespace std;
 
 class Client
