+++ Service.h
@@ -1,9 +1,5 @@
 #include <iostream>
-
-
 using namespace std;
-
-
 
 class Service
 {
